+++ src/js/load.js
@@ -1,3 +1,23 @@
+  prototype.init = function () {
+    var $this = this.$element,
+        url;
+
+    if ($this.is('img')) {
+      this.isImg = true;
+      this.originalUrl = url = $this.attr('src'); // e.g.: "img/picture.jpg"
+
+      if (!url) { // Blank image
+        return;
+      }
+
+      url = $this.prop('src'); // e.g.: "http://example.com/img/picture.jpg"
+    } else if ($this.is('canvas') && SUPPORT_CANVAS) {
+      url = $this[0].toDataURL();
+    }
+
+    this.load(url);
+  };
+
   prototype.load = function (url) {
     var options = this.options,
         $this = this.$element,
@@ -5,18 +25,6 @@
         bustCacheUrl,
         buildEvent,
         $clone;
-
-    if (!url) {
-      if ($this.is('img')) {
-        if (!$this.attr('src')) {
-          return;
-        }
-
-        url = $this.prop('src');
-      } else if ($this.is('canvas') && SUPPORT_CANVAS) {
-        url = $this[0].toDataURL();
-      }
-    }
 
     if (!url) {
       return;
@@ -30,18 +38,20 @@
     }
 
     if (options.checkImageOrigin && isCrossOriginURL(url)) {
-      crossOrigin = 'anonymous';
+      crossOrigin = ' crossOrigin="anonymous"';
 
       if (!$this.prop('crossOrigin')) { // Only when there was not a "crossOrigin" property
         bustCacheUrl = addTimestamp(url); // Bust cache (#148)
       }
     }
 
-    this.$clone = $clone = $('<img>');
+    // IE8 compatibility: Don't use "$().attr()" to set "src"
+    this.$clone = $clone = $('<img' + (crossOrigin || '') + ' src="' + (bustCacheUrl || url) + '">');
 
     $clone.one('load', $.proxy(function () {
-      var naturalWidth = $clone.prop('naturalWidth') || $clone.width(),
-          naturalHeight = $clone.prop('naturalHeight') || $clone.height();
+      var image = $clone[0],
+          naturalWidth = image.naturalWidth || image.width,
+          naturalHeight = image.naturalHeight || image.height; // $clone.width() and $clone.height() will return 0 in IE8 (#319)
 
       this.image = {
         naturalWidth: naturalWidth,
@@ -55,9 +65,6 @@
       this.build();
     }, this)).one('error', function () {
       $clone.remove();
-    }).attr({
-      crossOrigin: crossOrigin, // "crossOrigin" must before "src" (#271)
-      src: bustCacheUrl || url
     });
 
     // Hide and insert into the document